332 configurable datastore parameters (#349)
* added client_reader_to_csv example

* Revert "added client_reader_to_csv example"

This reverts commit 474d74567d5a4ee3ccbc94b2a7c7d8197c95e323.

* ConfigurationManager changes

* added client_reader_to_csv example

* Revert "added client_reader_to_csv example"

This reverts commit 474d74567d5a4ee3ccbc94b2a7c7d8197c95e323.

* make StoryPipeline intervals configurable

* make inactive_pipeline_delay configurable

* Refactoring ConfigurationManager

* changes in LogConf

* replaced ChronoLog namespace with chronolog

+++ chrono_common/StoryPipeline.h
@@ -22,8 +22,8 @@ class StoryPipeline
 
 public:
     StoryPipeline(StoryChunkExtractionQueue &, ChronicleName const &chronicle_name, StoryName const &story_name
-                  , StoryId const &story_id, uint64_t start_time, uint16_t chunk_granularity = 60 // seconds
-                  , uint16_t acceptance_window = 120 // seconds
+                  , StoryId const &story_id, uint64_t start_time, uint32_t chunk_granularity = 60 // seconds
+                  , uint32_t acceptance_window = 120 // seconds
     );
 
     StoryPipeline(StoryPipeline const &) = delete;
